+++ workflows/__init__.py
@@ -1,0 +1,1 @@
+"""CTI Scraper workflows package."""